fixed inline bug from clang complier; add require m_time module for lua.

+++ Makefile
@@ -1,15 +1,26 @@
-CC = gcc 
+CC = gcc
 CFLAGS += -O2  -g -Wall
 AR = ar rcu
-ifeq ($(OS), Windows_NT)
- N = \\
- RM = del
- CFLAGS += -D __USE_MINGW_ANSI_STDIO -D _MINGW32_
- A2 = a2.lib
-else
- N = //
- RM = rm -rf
- A2 = liba2.a
+
+
+ifeq ($(OS),Windows_NT)
+	N = \\
+	RM = del
+	CFLAGS += -D __USE_MINGW_ANSI_STDIO -D _MINGW32_
+	A2 = a2.lib 
+else 
+	UNAME_S := $(shell uname -s)
+	ifeq ($(UNAME_S), Darwin)
+		N = //
+		RM = rm -rf
+		A2 = liba2.a
+		CFLAGS += -std=gnu89
+	endif 
+	ifeq ($(UNAME_S), Linux)
+		N = //
+		RM = rm -rf
+		A2 = liba2.a
+	endif
 endif
 
 A2_OBJ = a2_mem.o a2_error.o a2_io.o a2_lex.o a2_map.o a2_string.o a2_env.o  a2_closure.o \
